@@ -2,7 +2,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.winterhaven-mc</groupId>
 	<artifactId>SpawnStar</artifactId>
-	<version>1.18.3</version>
+	<version>1.18.4</version>
 	<name>SpawnStar</name>
 
 	<url>https://github.com/tim-savage/SpawnStar</url>
@@ -15,6 +15,7 @@
 	<properties>
 		<project.author>Tim Savage</project.author>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<maven.compiler.release>11</maven.compiler.release>
 	</properties>
 
 	<licenses>
@@ -70,25 +71,25 @@
 			<version>2.2.1</version>
 		</dependency>
 
-		<!-- MessengerLib -->
+		<!-- MessageBuilderLib -->
 		<dependency>
 			<groupId>com.winterhaven-mc</groupId>
 			<artifactId>message-builder-lib</artifactId>
-			<version>1.0.4</version>
+			<version>1.11.1</version>
 		</dependency>
 
 		<!-- SoundConfigLib -->
 		<dependency>
 			<groupId>com.winterhaven-mc</groupId>
 			<artifactId>sound-config-lib</artifactId>
-			<version>1.16.3</version>
+			<version>1.11.0</version>
 		</dependency>
 
 		<!-- WorldManagerLib -->
 		<dependency>
 			<groupId>com.winterhaven-mc</groupId>
 			<artifactId>world-manager-lib</artifactId>
-			<version>1.16.3</version>
+			<version>1.11.0</version>
 		</dependency>
 
 		<!-- Unit Testing Dependencies -->
@@ -157,11 +158,7 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.8.1</version>
-				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
-				</configuration>
+				<version>3.9.0</version>
 			</plugin>
 
 			<!-- Create Shaded jar -->
@@ -226,7 +223,6 @@
 				<artifactId>maven-javadoc-plugin</artifactId>
 				<version>3.3.1</version>
 				<configuration>
-					<source>8</source>
 					<doclint>all,-missing</doclint>
 				</configuration>
 				<executions>
