@@ -32,11 +32,10 @@ public final class SpawnStarFactory {
 	private final List<String> itemStackLore;
 
 	// item metadata flags
-	private static final Set<ItemFlag> itemFlagSet =
-			Collections.unmodifiableSet(new HashSet<>(Arrays.asList(
+	private static final Set<ItemFlag> itemFlagSet = Set.of(
 					ItemFlag.HIDE_ATTRIBUTES,
 					ItemFlag.HIDE_ENCHANTS,
-					ItemFlag.HIDE_UNBREAKABLE)));
+					ItemFlag.HIDE_UNBREAKABLE );
 
 	// the proto item
 	private final ItemStack protoItem;
