@@ -1,7 +1,9 @@
 package com.winterhavenmc.spawnstar.commands;
 
 import com.winterhavenmc.spawnstar.PluginMain;
+import com.winterhavenmc.spawnstar.messages.MessageId;
 import com.winterhavenmc.spawnstar.sounds.SoundId;
+
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 
@@ -9,17 +11,15 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static com.winterhavenmc.spawnstar.messages.MessageId.*;
-import static com.winterhavenmc.spawnstar.sounds.SoundId.*;
-
 
 /**
  * Help command implementation<br>
  * displays help and usage messages for plugin commands
  */
-public class HelpCommand extends AbstractSubcommand implements Subcommand {
+final class HelpCommand extends SubcommandAbstract implements Subcommand {
 
 	private final PluginMain plugin;
+	private final SubcommandRegistry subcommandRegistry;
 
 
 	/**
@@ -27,11 +27,12 @@ public class HelpCommand extends AbstractSubcommand implements Subcommand {
 	 *
 	 * @param plugin reference to plugin main class instance
 	 */
-	HelpCommand(final PluginMain plugin) {
+	HelpCommand(final PluginMain plugin, SubcommandRegistry subcommandRegistry) {
 		this.plugin = Objects.requireNonNull(plugin);
+		this.subcommandRegistry = subcommandRegistry;
 		this.name = "help";
 		this.usage = "/spawnstar help [command]";
-		this.description = COMMAND_HELP_HELP;
+		this.description = MessageId.COMMAND_HELP_HELP;
 		this.permission = "spawnstar.help";
 		this.maxArgs = 1;
 	}
@@ -45,7 +46,7 @@ public class HelpCommand extends AbstractSubcommand implements Subcommand {
 
 		if (args.length == 2) {
 			if (args[0].equalsIgnoreCase("help")) {
-				for (String subcommand : subcommandMap.getKeys()) {
+				for (String subcommand : subcommandRegistry.getKeys()) {
 					if (sender.hasPermission(permission)
 							&& subcommand.startsWith(args[1].toLowerCase())
 							&& !subcommand.equalsIgnoreCase("help")) {
@@ -64,14 +65,14 @@ public class HelpCommand extends AbstractSubcommand implements Subcommand {
 
 		// if command sender does not have permission to display help, output error message and return true
 		if (!sender.hasPermission(permission)) {
-			plugin.messageBuilder.build(sender, COMMAND_FAIL_HELP_PERMISSION).send();
+			plugin.messageBuilder.build(sender, MessageId.COMMAND_FAIL_HELP_PERMISSION).send();
 			plugin.soundConfig.playSound(sender, SoundId.COMMAND_FAIL);
 			return true;
 		}
 
 		// check max arguments
 		if (args.size() > getMaxArgs()) {
-			plugin.messageBuilder.build(sender, COMMAND_FAIL_ARGS_COUNT_OVER).send();
+			plugin.messageBuilder.build(sender, MessageId.COMMAND_FAIL_ARGS_COUNT_OVER).send();
 			plugin.soundConfig.playSound(sender, SoundId.COMMAND_FAIL);
 			displayUsage(sender);
 			return true;
@@ -99,7 +100,7 @@ public class HelpCommand extends AbstractSubcommand implements Subcommand {
 	void displayHelp(final CommandSender sender, final String commandName) {
 
 		// get subcommand from map by name
-		Subcommand subcommand = subcommandMap.getCommand(commandName);
+		Subcommand subcommand = subcommandRegistry.getCommand(commandName);
 
 		// if subcommand found in map, display help message and usage
 		if (subcommand != null) {
@@ -109,8 +110,8 @@ public class HelpCommand extends AbstractSubcommand implements Subcommand {
 
 		// else display invalid command help message and usage for all commands
 		else {
-			plugin.messageBuilder.build(sender, COMMAND_HELP_INVALID).send();
-			plugin.soundConfig.playSound(sender, COMMAND_INVALID);
+			plugin.messageBuilder.build(sender, MessageId.COMMAND_HELP_INVALID).send();
+			plugin.soundConfig.playSound(sender, SoundId.COMMAND_INVALID);
 			displayUsageAll(sender);
 		}
 	}
@@ -123,11 +124,11 @@ public class HelpCommand extends AbstractSubcommand implements Subcommand {
 	 */
 	void displayUsageAll(final CommandSender sender) {
 
-		plugin.messageBuilder.build(sender, COMMAND_HELP_USAGE_HEADER).send();
+		plugin.messageBuilder.build(sender, MessageId.COMMAND_HELP_USAGE_HEADER).send();
 
-		for (String subcommandName : subcommandMap.getKeys()) {
-			if (subcommandMap.getCommand(subcommandName) != null) {
-				subcommandMap.getCommand(subcommandName).displayUsage(sender);
+		for (String subcommandName : subcommandRegistry.getKeys()) {
+			if (subcommandRegistry.getCommand(subcommandName) != null) {
+				subcommandRegistry.getCommand(subcommandName).displayUsage(sender);
 			}
 		}
 	}
